msemblator/requirements.txt
numpy==1.26.4
pandas==2.1.4
rdkit==2024.9.6
scikit-learn==1.2.2
joblib==1.2.0
tqdm==4.65.0
pyyaml==6.0.1
wexpect==4.0.0
lightgbm==3.3.5
msbuddy==0.3.8